move test optimization integration tests to dedicated folder
