+++ .github/workflows/test-optimization.yml
@@ -34,6 +34,19 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ steps.app-token.outputs.token }}
 
+  integration:
+    strategy:
+      matrix:
+        version: [oldest, maintenance, active, latest]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+      - uses: ./.github/actions/node
+        with:
+          version: ${{ matrix.version }}
+      - uses: ./.github/actions/install
+      - run: yarn test:integration:testopt
+
   integration-playwright:
     strategy:
       matrix:
